Finished Scores, updated main, recorded sounds
completed the scores file which now can get the score, start the driver and close the driver in seperate functions as well as check if the game is live. The main functionworks with scores functions and will end when the game finishes. The sounds were recorded and implemented in mp3 format.